Don't needlessly generate a bunch of svgs for the lines.
Instead we just generate one, and use slint to repeat it across the width of
the window. This still needs some more work because the height of the squiggle
svg is hardcode to 1000, so we need to extract that out somehow and link it
to the height of the window? Or maybe we just need to repeat it vertically somehow too.